dernier commit

+++ src/main/java/com/sqli/gestiondeproduits/service/ProduitServiceImpl.java
@@ -2,6 +2,7 @@ package com.sqli.gestiondeproduits.service;
 
 import com.sqli.gestiondeproduits.Exception.DuplicateProductException;
 import com.sqli.gestiondeproduits.Exception.PrixProduitException;
+import com.sqli.gestiondeproduits.Exception.ProduitNExistePas;
 import com.sqli.gestiondeproduits.dao.ProduitDao;
 import com.sqli.gestiondeproduits.model.Produit;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,7 +17,7 @@ public class ProduitServiceImpl implements ProduitService{
     @Autowired
     private  ProduitDao produitDao;
     @Override
-    public Optional<List<Produit>> getAllProduits() {
+    public Optional<List<Produit>> recupererTousLesProduits() {
         try {
             List<Produit> produits= produitDao.findAll();
             return Optional.ofNullable(produits);
@@ -28,12 +29,18 @@ public class ProduitServiceImpl implements ProduitService{
     }
 
     @Override
-    public Optional<Produit> getProduitById(Integer id) {
+    public Optional<Produit> recupererProduitById(Integer id) {
+        if(produitDao.findById(id).isEmpty()){
+            throw new ProduitNExistePas("le produit que vous voulez supprimer n'exitse pas ");
+        }
         return produitDao.findById(id);
     }
 
     @Override
     public void ajouterProduit(Produit produit) {
+        if(produit.getPrixProduit()<=0){
+            throw  new PrixProduitException("le prix d'un produit doit etre superieur a 0");
+        }
         if (produitDao.findByCodeProduit(produit.getCodeProduit()).isPresent()){
             throw new DuplicateProductException("un produit avec le meme code deja existe");
         }
@@ -43,7 +50,7 @@ public class ProduitServiceImpl implements ProduitService{
     }
 
     @Override
-    public void updateProduit(Integer idProduit, Produit produit) {
+    public void mettreAJourProduit(Integer idProduit, Produit produit) {
         if(produit.getPrixProduit()<=0){
             throw  new PrixProduitException("le prix d'un produit doit etre superieur a 0");
         }
@@ -60,7 +67,10 @@ public class ProduitServiceImpl implements ProduitService{
     }
 
     @Override
-    public void deleteProduit(Integer idProduit) {
+    public void supprimerProduit(Integer idProduit) {
+        if(produitDao.findById(idProduit).isEmpty()){
+            throw new ProduitNExistePas("le produit que vous voulez supprimer n'exitse pas ");
+        }
         produitDao.deleteById(idProduit);
     }
 }
